refactor(network): split mega lib module
